Improve map drawer search and results presentation.
This PR improves the map drawer search by adding more fields to be included in the fuzzy search such as directory and subjects. I extracted the search logic into a standalone util that can be used elsewhere if needed.

Lastly, I added director to the result card used to display search results in the map drawer.
