+++ components/result-card.tsx
@@ -22,14 +22,16 @@ export function ResultCard({ media, isSelected }: ResultCardProps) {
       aria-label={`Select ${media.name}`}
     >
       <div className="flex items-start justify-between gap-2">
-        <div className="min-w-0">
+        <div className="min-w-0 space-y-2">
           <p className="font-medium text-sm truncate">
             {media.name}{" "}
             {media.media?.release_year && `(${media.media.release_year})`}
           </p>
-          {media.country && (
+          {(media.country || media.media?.director) && (
             <p className="text-xs text-muted-foreground truncate">
-              {media.country}
+              {[media.country, media.media?.director]
+                .filter(Boolean)
+                .join(" · ")}
             </p>
           )}
         </div>
